refactor(constants): 숫자 상수 이름 변경
